+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Localized the quotation marks on the Notifications view.
 - Fixed a bug where nostr entities in URLs were treated like quoted note links.
 - Added in-app profile photo editing.
+- Changed "Name" to "Display Name" on the Edit Profile View.
 
 ### Internal Changes
 - Included the npub in the properties list sent to analytics.